feat: Update layouts (#16)
