+++ src/app/app.routes.ts
@@ -2,9 +2,9 @@ import {Routes} from '@angular/router';
 import {RegisterComponent} from './pages/register/register.component';
 import {LoginComponent} from './pages/login/login.component';
 import {HomepageComponent} from './pages/home/homepage.component';
-import {GamePageComponent} from './pages/game-page/game-page.component';
 import {ProfileComponent} from './pages/profile/profile.component';
 import {RankingComponent} from './pages/ranking/ranking.component';
+import {GameComponent} from './pages/game/game.component';
 
 export const routes: Routes = [
   {
@@ -25,7 +25,7 @@ export const routes: Routes = [
   },
   {
     path: "game/:action",
-    component: GamePageComponent
+    component: GameComponent
   },
   {
     path: "profile",
