+++ src/app/components/default-dashboard-layout/sidenav/sidenav.component.html
@@ -9,7 +9,7 @@
   <!-- User Info Section -->
   <div class="flex flex-col w-full items-center gap-3 backdrop-blur-xl  rounded-lg p-2">
     <div class="overflow-hidden rounded-full w-28 h-28">
-      <img src="assets/illustration-game-card.png" alt="illustration-game-card">
+      <img class="w-full" src="assets/user-profile.png" alt="illustration-game-card">
     </div>
 
     <div>
