+++ src/app/pages/home/homepage.component.html
@@ -27,7 +27,7 @@
 
             <!-- Imagem principal -->
             <img class="relative w-40 h-auto "
-                 src="/assets/illustration-game-card-blue_5BJXOdF.png"
+                 src="/assets/illustration-game-card.png"
                  alt="{{ data.game_title }}-illustration">
           </div>
