+++ src/app/pages/login/login.component.html
@@ -14,7 +14,7 @@
       formControlName="username"
       type=text
       label="Usename"
-      placeholder="username"
+      placeholder="Username"
 
     >
     </app-primary-input>
@@ -28,7 +28,7 @@
       formControlName="password"
       type="password"
       label="Password"
-      placeholder="Enter yout password..."
+      placeholder="Enter your password..."
     >
     </app-primary-input>
